requirements.txt
flask
flask-cors
openai
psycopg2-binary
python-dotenv
langgraph
langchain-openai
langchain-core
